+++ external/llvm-project/mlir/include/mlir/Conversion/TosaToTensor/TosaToTensor.h
@@ -13,6 +13,7 @@
 #ifndef MLIR_CONVERSION_TOSATOTENSOR_TOSATOTENSOR_H
 #define MLIR_CONVERSION_TOSATOTENSOR_TOSATOTENSOR_H
 
+#include "mlir/IR/PatternMatch.h"
 #include "mlir/Pass/Pass.h"
 
 namespace mlir {
